+++ scripts/com.r2studio.TsumBeta/src/index.js
@@ -181,6 +181,20 @@ var AbstractPage = {
       {x: 760, y: 760, r: 247, g: 182, b: 8, match: true, threshold: 30}, // yellow button premium box top left
       {x: 960, y: 980, r: 247, g: 178, b: 8, match: true, threshold: 30}  // right side of premium box button
     ],
+    colors2025OverviewGeneral: [
+      {x: 30, y: 910, r: 16, g: 53, b: 93, match: true, threshold: 30},
+      {x: 60, y: 910, r: 233, g: 171, b: 8, match: true, threshold: 30},
+      {x: 520, y: 910, r: 237, g: 174, b: 8, match: true, threshold: 30},
+      {x: 545, y: 840, r: 22, g: 65, b: 107, match: true, threshold: 30},
+      {x: 570, y: 910, r: 29, g: 85, b: 159, match: true, threshold: 30},
+      {x: 10, y: 955, r: 37, g: 71, b: 115, match: true, threshold: 30},
+      {x: 170, y: 1490, r: 48, g: 81, b: 130, match: true, threshold: 30},
+      {x: 170, y: 1515, r: 8, g: 164, b: 213, match: true, threshold: 30},
+      {x: 170, y: 1570, r: 247, g: 194, b: 16, match: true, threshold: 30}
+    ],
+    colors2025OverviewSelected: [
+      {x: 440, y: 1200, r: 246, g: 181, b: 17, match: true, threshold: 20} // yellow area for either premium or special box
+    ],
     colorsNoMissionPage: [
       {x: 170, y: 1580, r: 49, g: 40, b: 0, match: true, threshold: 30}   // left bottom modal button
     ],
@@ -430,6 +444,20 @@ var Page = {
     next: {x: 176, y: 1592},
     store: {x: 910, y: 1592}
   },
+  TsumTsum2025StorePage: {
+    // needs to be before PrePage as it is more specific than the other page
+    name: 'TsumTsumStorePage',
+    colors: AbstractPage.TsumStore.colors2025OverviewGeneral.concat(AbstractPage.TsumStore.colors2025OverviewSelected),
+    back: {x: 190, y: 1650},
+    next: {x: 910, y: 690}
+  },
+  TsumTsum2025StorePrePage: {
+    name: 'TsumTsumStorePrePage',
+    colors: AbstractPage.TsumStore.colors2025OverviewGeneral,
+    back: {x: 190, y: 1650},
+    store: {x: 430, y: 1200},
+    next: {x: 430, y: 1200}
+  },
   TsumTsumStorePageTwoBoxesNoMission: {
     name: 'TsumTsumStorePage',
     colors: AbstractPage.TsumStore.colorsTwoBoxTypes.concat(AbstractPage.TsumStore.colorsNoMissionPage),
@@ -459,11 +487,25 @@ var Page = {
     colors: [
       {x: 208, y: 1070, r: 247, g: 176, b: 8, match: true, threshold: 30},  // left of Cancel button
       {x: 420, y: 1070, r: 247, g: 176, b: 8, match: true, threshold: 30},  // right of Cancel button
-      {x: 540, y: 1070, r: 54, g: 93, b: 146, match: true, threshold: 30},  // betwen buttons
+      {x: 540, y: 1070, r: 54, g: 93, b: 146, match: true, threshold: 30},  // between buttons
       {x: 650, y: 1070, r: 247, g: 176, b: 8, match: true, threshold: 30},  // left of OK button
       {x: 880, y: 1070, r: 247, g: 176, b: 8, match: true, threshold: 30},  // right of OK button
       {x: 948, y: 1066, r: 33, g: 69, b: 107, match: true, threshold: 30},  // right next to OK button
       {x: 805, y: 1265, r: 239, g: 167, b: 8, match: true, threshold: 50}   // left of List button
+    ],
+    back: {x: 310, y: 1070},  // Cancel button
+    next: {x: 760, y: 1070}   // OK button
+  },
+  Confirm2025PurchaseBoxPage: {
+    name: 'ConfirmPurchasePage',
+    colors: [
+      {x: 208, y: 1070, r: 247, g: 186, b:   8, match: true, threshold: 30},  // left of Cancel button
+      {x: 420, y: 1070, r: 247, g: 184, b:   8, match: true, threshold: 30},  // right of Cancel button
+      {x: 540, y: 1070, r:  54, g:  90, b: 141, match: true, threshold: 30},  // between buttons
+      {x: 650, y: 1070, r: 247, g: 190, b:   8, match: true, threshold: 30},  // left of OK button
+      {x: 880, y: 1070, r: 247, g: 191, b:  14, match: true, threshold: 30},  // right of OK button
+      {x: 948, y: 1066, r:  40, g:  70, b: 113, match: true, threshold: 30},  // right next to OK button
+      {x: 785, y: 1320, r: 238, g: 171, b:   8, match: true, threshold: 50}   // left of List button
     ],
     back: {x: 310, y: 1070},  // Cancel button
     next: {x: 760, y: 1070}   // OK button
@@ -473,7 +515,7 @@ var Page = {
     colors: [
       {x: 200, y: 1444, r: 247, g: 178, b: 8, match: true, threshold: 30},  // left of Cancel button
       {x: 426, y: 1444, r: 247, g: 178, b: 8, match: true, threshold: 30},  // right of Cancel button
-      {x: 540, y: 1444, r: 54, g: 93, b: 146, match: true, threshold: 30},  // betwen buttons
+      {x: 540, y: 1444, r: 54, g: 93, b: 146, match: true, threshold: 30},  // between buttons
       {x: 660, y: 1444, r: 247, g: 174, b: 8, match: true, threshold: 30},  // left of OK button
       {x: 860, y: 1444, r: 247, g: 178, b: 8, match: true, threshold: 30},  // right of OK button
       {x: 940, y: 1444, r: 33, g: 65, b: 107, match: true, threshold: 30},  // right next to OK button
@@ -1719,15 +1761,17 @@ Tsum.prototype.goTsumTsumStorePage = function() {
       this.startApp();
     }
     this.goTsumsPage();
-    this.tap(Page.TsumsPage.store);
-    this.sleep(3000);
-    var pageName = this.findPage(2, 2000);
-    if (pageName === 'TsumTsumStorePage') {
-      return true;
-    } else {
-      log('Unexpected page found:', pageName, 'goTsumTsumStorePage');
-      return false;
+    for (var i = 0; i < 3; i++) {
+      this.tap(this.findPageObject().store);
+      this.sleep(3000);
+      var pageName = this.findPage(2, 2000);
+      log("Pg: ", pageName);
+      if (pageName === 'TsumTsumStorePage') {
+        return true;
+      }
     }
+    log('Unexpected page found:', pageName, 'goTsumTsumStorePage');
+    return false;
   }
 }
 
@@ -2754,14 +2798,20 @@ Tsum.prototype.taskAutoBuyBoxes = function() {
   }
   this.lastVisitedPages.autoBuyBoxesStore = true;
   log("Start buying ", this.autobuyBoxes, "boxes - taskAutoBuyBoxes");
-  var countUnknownPages = 0;
-  while (this.isRunning && this.autobuyBoxes > 0) {
+  var countUnknownPages = 0, countSamePage = 0;
+  while (this.isRunning && this.autobuyBoxes > 0 && countSamePage < 10) {
     this.requestTsumMonitor();
     var page = this.findPageObject(1, 200);
     if (page != null) {
       countUnknownPages = 0;
       this.lastVisitedPages['autoBuyBoxes' + page.name] = true;
       this.tap(page.next);
+      if (page === lastPage) {
+        countSamePage++;
+        debug("countSamePage = ", countSamePage);
+      } else {
+        countSamePage = 0;
+      }
       if (page !== lastPage && page === Page.BoxPurchasedPage) {
         this.autobuyBoxes--;
         log("Bought box.", this.autobuyBoxes, "left");